requirements.txt
numpy==1.26.1
opencv-python==4.8.1.78
Pillow==10.1.0
requests==2.31.0